Refatoração no atendimento. Novo arquivo para armazenar funções do formulário.
