+++ app/controller/primeiroAtendimento.py
@@ -1,21 +1,13 @@
-from dao.paciente import Paciente
 from datetime import datetime
-from dao.atendimento import AtendimentoBuilder, inserirPaciente
+from dao.atendimento import AtendimentoBuilder
+from dao.paciente import inserirPaciente
 from flask_login import current_user
-import re
-
-
-# +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ #
-# O QUE FALTA: 
-# 1- Adicionar doenças cronicas/medicamentos
-# 2- Informações dos fieldsets 5 e 6#
+from controller.formfuncs import *
 
 
 def registrar(form):
     # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ #
-    # Para testes: essas informações precisam vir por parâmetro
     data = datetime.today()
-    # data = datetime.strptime(data, '%d/%m/%Y').date() if len(data) != 0 else None
     # ============== Paciente ==============
 
     nome = data_or_null(form['nome'])
@@ -288,7 +280,6 @@ def registrar(form):
 
         print('has_sintomas: {}'.format(has_sintomas))
 
-        #VERIFICAR COMO ESSAS INFOS ESTÃO VINDO PARA CADASTRAR
         if has_sintomas == 1:  # Sim
 
             size = data_or_null(form['has_sintoma_len'], int)
@@ -326,48 +317,4 @@ def registrar(form):
     # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ #
 
 
-def get_real_data(data: list) -> list:
-    return list(map(format_real_data, filter(filter_real_data, data)))
 
-
-def get_others_data(data: list) -> list:
-    return list(filter(filter_others_data, data))
-
-
-def filter_real_data(data: str) -> bool:
-    return 'real_data_' in data
-
-
-def filter_others_data(data: str) -> bool:
-    return not filter_real_data(data)
-
-
-def format_real_data(data: str) -> int:
-    return int(data.replace('real_data_', ''))
-
-
-def data_or_null(data: str, cast=None):
-    if cast is None:
-        cast = str
-
-    return cast(data) if len(data) != 0 else None
-
-
-def only_num(data: str):
-    if data is None:
-        return None
-
-    return re.sub('[^\\d]', '', data)
-
-
-def multiselect(form, name: str, size) -> list:
-    ret = []
-
-    for i in range(1, int(size) + 1):
-        name = '{}_{}'.format(name, i).replace('_1', '')
-        if name in form:
-            ret.append(data_or_null(form[name]))
-        else:
-            ret.append(None)
-
-    return ret
